chore(web): update tracked changes
- ConfirmDialog
- Simplified component logic
- Reused ConfirmDialog for consistency
- Removed redundant state and effects
- Updated styles and text content
