+++ web/src/lib/components/ui/index.ts
@@ -2,3 +2,4 @@ export { default as SkeletonText } from "./SkeletonText.svelte";
 export { default as SkeletonThumbnail } from "./SkeletonThumbnail.svelte";
 export { default as SkeletonVideoList } from "./SkeletonVideoList.svelte";
 export { default as SkeletonMediaList } from "./SkeletonMediaList.svelte";
+export { default as ConfirmDialog } from "./ConfirmDialog.svelte";
